fix(fscomponents): cross platform imports
