+++ packages/fscomponents/package.json
@@ -36,6 +36,7 @@
     "react-native-picker-select": "^8.0.0",
     "react-native-svg": "^10.0.0",
     "react-native-swiper": "1.6.0",
+    "react-native-vector-icons": "^8.0.0",
     "react-native-video": "^5.0.2",
     "react-native-web-modal": "^1.0.1",
     "react-native-webview": "^10.9.0",
