+++ packages/fscomponents/src/components/serializable/VideoBlock.tsx
@@ -1,47 +1,16 @@
+
+import type { StateType, VideoBlockProps } from './VideoBlock.types';
+
 import React, { Component } from 'react';
 import {
   Dimensions,
   StyleSheet,
   TouchableOpacity,
-  View,
-  ViewStyle
+  View
 } from 'react-native';
 import WebView from 'react-native-webview';
-import VideoPlayer, { VideoProperties } from 'react-native-video';
-import { ReactPlayerProps } from 'react-player';
-import { VideoProgressStateWeb } from './VideoBlock.web';
+import VideoPlayer from 'react-native-video';
 import { cloneDeep } from 'lodash-es';
-
-export interface VideoProgressStateNative {
-  currentTime: number;
-  playableDuration: number;
-  seekableDuration: number;
-}
-
-export interface VideoSource {
-  uri: string;
-  ratio?: number;
-}
-
-export interface VideoBlockProps extends ReactPlayerProps,
-  Omit<VideoProperties, 'style' | 'onBuffer'> {
-  source: VideoSource;
-  autoPlay?: boolean;
-  repeat?: boolean;
-  resizeMode?: 'stretch' | 'contain' | 'cover' | 'none';
-  style?: ViewStyle | ViewStyle[];
-  muted?: boolean;
-  fullscreen?: boolean;
-  onError?: (error: any, data?: any, hlsInstance?: any, hlsGlobal?: any) => void;
-  onSeek?: (data?: any) => void;
-  onProgress?: (data: VideoProgressStateNative | VideoProgressStateWeb) =>
-    void;
-}
-
-export interface StateType {
-  videoPaused: boolean;
-  facebookSrc: string;
-}
 
 const styles = StyleSheet.create({
   VideoButton: {
