+++ packages/fscomponents/src/components/serializable/index.ts
@@ -6,5 +6,4 @@ export * from './SerializableShareButton';
 export * from './SerializableButton';
 export * from './SerializableImage';
 export * from './SerializableImagePlaceholder';
-export * from './VideoBlock.native';
-export * from './VideoBlock.web';
+export * from './VideoBlock';
